@@ -41,65 +41,61 @@
   },
   "homepage": "https://github.com/ForgeRock/forgerock-javascript-sdk#readme",
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.6",
-    "@angular-devkit/core": "^13.2.6",
-    "@nrwl/angular": "^13.8.8",
-    "@nrwl/cli": "13.8.8",
-    "@nrwl/eslint-plugin-nx": "13.8.8",
-    "@nrwl/express": "13.8.8",
-    "@nrwl/jest": "13.8.8",
-    "@nrwl/js": "13.8.8",
-    "@nrwl/linter": "13.8.8",
-    "@nrwl/node": "13.8.8",
-    "@nrwl/react": "13.8.8",
-    "@nrwl/schematics": "^8.12.11",
-    "@nrwl/tao": "^13.3.9",
-    "@nrwl/web": "^13.3.9",
-    "@nrwl/workspace": "^13.3.9",
-    "@playwright/test": "^1.17.1",
-    "@swc/core": "^1.2.126",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/react-hooks": "7.0.2",
-    "@types/express": "^4.17.0",
-    "@types/jest": "^27.0.0",
-    "@types/jest-dev-server": "^5.0.0",
-    "@types/react": "17.0.30",
-    "@types/react-dom": "17.0.9",
-    "@typescript-eslint/eslint-plugin": "5.3.1",
-    "@typescript-eslint/parser": "5.3.1",
-    "@typescript-eslint/typescript-estree": "^4.7.0",
-    "babel-jest": "27.2.3",
-    "concurrently": "^6.4.0",
-    "eslint": "8.2.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "2.25.2",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "7.27.1",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "html-webpack-plugin": "^4.5.2",
-    "husky": "^4.3.8",
-    "jest": "^27.3.1",
-    "jest-canvas-mock": "^2.2.0",
-    "jest-environment-jsdom": "^27.3.1",
-    "jest-junit": "^12.0.0",
-    "jest-preset-angular": "11.0.0",
+    "@nrwl/angular": "14.5.7",
+    "@nrwl/cli": "14.5.7",
+    "@nrwl/eslint-plugin-nx": "14.5.7",
+    "@nrwl/express": "14.5.7",
+    "@nrwl/jest": "14.5.7",
+    "@nrwl/js": "14.5.7",
+    "@nrwl/linter": "14.5.7",
+    "@nrwl/node": "14.5.7",
+    "@nrwl/react": "14.5.7",
+    "@nrwl/web": "14.5.7",
+    "@nrwl/workspace": "14.5.7",
+    "@playwright/test": "1.17.1",
+    "@swc/core": "1.2.241",
+    "@types/express": "4.17.0",
+    "@types/jest": "27.4.1",
+    "@types/jest-dev-server": "5.0.0",
+    "@types/react": "18.0.17",
+    "@types/react-dom": "18.0.6",
+    "@typescript-eslint/eslint-plugin": "5.33.1",
+    "@typescript-eslint/parser": "5.33.1",
+    "@typescript-eslint/typescript-estree": "4.7.0",
+    "babel-jest": "27.5.1",
+    "concurrently": "6.4.0",
+    "core-js": "3.19.2",
+    "eslint": "8.15.0",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-prettier": "3.1.1",
+    "eslint-plugin-react": "7.30.1",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "html-webpack-plugin": "4.5.2",
+    "husky": "4.3.8",
+    "jest": "27.5.1",
+    "jest-canvas-mock": "2.2.0",
+    "jest-environment-jsdom": "27.3.1",
+    "jest-junit": "12.0.0",
+    "jest-preset-angular": "11.1.2",
     "local-web-server": "4.2.1",
-    "ngx-deploy-npm": "^3.0.5",
-    "playwright": "^1.17.1",
-    "prettier": "^2.1.1",
-    "pretty-quick": "^3.1.2",
-    "react-test-renderer": "17.0.2",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-visualizer": "^5.5.2",
-    "shx": "^0.3.2",
-    "swc-loader": "^0.1.15",
-    "ts-jest": "^27.0.7",
-    "ts-loader": "^8.0.17",
-    "typedoc": "^0.17.7",
-    "typescript": "^4.0.0",
-    "webpack": "^5.64.1",
-    "webpack-cli": "^4.7.2"
+    "nx": "14.5.7",
+    "playwright": "1.17.1",
+    "prettier": "2.7.1",
+    "pretty-quick": "3.1.2",
+    "regenerator-runtime": "0.13.7",
+    "rollup-plugin-terser": "7.0.2",
+    "rollup-plugin-visualizer": "5.5.2",
+    "shx": "0.3.2",
+    "swc-loader": "0.1.15",
+    "ts-jest": "27.1.4",
+    "ts-loader": "8.0.17",
+    "ts-node": "10.9.1",
+    "typedoc": "0.17.7",
+    "typescript": "4.7.4",
+    "webpack": "5.64.1",
+    "webpack-cli": "4.7.2"
   },
   "workspaces": [
     "samples/*",
