@@ -13,25 +13,27 @@
   "author": "David Adams <david.adams@forgerock.com>",
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^13.2.6",
-    "@angular/compiler": "~13.2.6",
-    "@angular/animations": "~13.2.6",
-    "@angular/common": "~13.2.6",
-    "@angular/core": "~13.2.6",
-    "@angular/forms": "~13.2.6",
-    "@angular/platform-browser": "~13.2.6",
-    "@angular/platform-browser-dynamic": "~13.2.6",
-    "@angular/router": "~13.2.6",
-    "bootstrap": "^5.1.3",
-    "zone.js": "~0.11.4"
+    "@angular-devkit/core": "14.1.3",
+    "@angular/compiler": "14.1.3",
+    "@angular/animations": "14.1.3",
+    "@angular/common": "14.1.3",
+    "@angular/core": "14.1.3",
+    "@angular/forms": "14.1.3",
+    "@angular/platform-browser": "14.1.3",
+    "@angular/platform-browser-dynamic": "14.1.3",
+    "@angular/router": "14.1.3",
+    "bootstrap": "5.1.3",
+    "zone.js": "0.11.8"
   },
   "devDependencies": {
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.2.6",
-    "@angular/compiler-cli": "~13.2.6",
-    "@angular/language-service": "~13.2.6",
-    "@types/node": "^12.11.1"
+    "@angular-devkit/build-angular": "14.1.3",
+    "@angular-devkit/core": "14.1.3",
+    "@angular-eslint/eslint-plugin": "14.0.2",
+    "@angular-eslint/eslint-plugin-template": "14.0.2",
+    "@angular-eslint/template-parser": "14.0.2",
+    "@angular/cli": "14.1.3",
+    "@angular/compiler-cli": "14.1.3",
+    "@angular/language-service": "14.1.3",
+    "@types/node": "18.7.6"
   }
 }
