@@ -11,39 +11,29 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@babel/cli": "^7.14.5",
-    "@babel/core": "^7.14.6",
-    "@babel/eslint-parser": "^7.14.7",
-    "@babel/preset-env": "^7.14.5",
-    "@babel/preset-react": "^7.14.5",
-    "@nrwl/web": "^13.3.9",
-    "autoprefixer": "^10.2.6",
-    "babel-loader": "^8.2.2",
-    "bootstrap": "^5.0.1",
-    "css-loader": "^5.2.6",
-    "eslint": "8.2.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jsx": "^0.1.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "local-web-server": "^4.2.1",
-    "mini-css-extract-plugin": "^1.6.0",
-    "postcss": "^8.3.5",
-    "postcss-loader": "^6.1.0",
-    "prettier": "^2.3.2",
-    "sass": "^1.35.1",
-    "sass-loader": "^12.1.0",
+    "@babel/cli": "7.18.10",
+    "@babel/core": "7.18.13",
+    "@babel/eslint-parser": "7.18.9",
+    "@babel/preset-env": "7.18.10",
+    "@babel/preset-react": "7.18.6",
+    "autoprefixer": "10.4.8",
+    "babel-loader": "8.2.5",
+    "bootstrap": "5.2.0",
+    "css-loader": "6.7.1",
+    "eslint": "8.22.0",
+    "eslint-plugin-jsx": "0.1.0",
+    "mini-css-extract-plugin": "2.6.1",
+    "sass": "1.54.5",
     "shx": "^0.3.3",
-    "style-loader": "^2.0.0",
-    "webpack": "^5.39.0",
-    "webpack-cli": "^4.7.2"
+    "webpack": "5.74.0",
+    "webpack-cli": "4.10.0"
   },
   "dependencies": {
-    "dotenv": "^10.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0"
+    "dotenv": "10.0.0",
+    "react": "18.2.0",
+    "react-is": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.3.0"
   },
   "engines": {
     "node": ">=14.2.0"
