@@ -1,5 +1,6 @@
-module.exports = {
-  displayName: 'mock-server',
+/* eslint-disable */
+export default {
+  displayName: 'mock-api',
   preset: '../../jest.preset.js',
   globals: {
     'ts-jest': {
@@ -11,5 +12,5 @@ module.exports = {
     '^.+\\.[tj]s$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/samples/mock-server',
+  coverageDirectory: '../../coverage/samples/mock-api',
 };
