@@ -1,4 +1,5 @@
-module.exports = {
+/* eslint-disable */
+export default {
   displayName: '..-samples--react-js',
   preset: '../../jest.preset.js',
   transform: {
